@@ -38,12 +38,12 @@
     "promise": "8.0.1",
     "prop-types": "^15.6.1",
     "raf": "3.4.0",
-    "react": "^16.3.1",
+    "react": "^16.12.0",
     "react-bootstrap": "^0.32.1",
     "react-context-component": "^0.0.6",
     "react-dev-utils": "^5.0.0",
-    "react-dom": "^16.2.0",
-    "react-redux": "^5.0.7",
+    "react-dom": "^16.12.0",
+    "react-redux": "^7.1.3",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "react-test-utils": "^0.0.1",
