@@ -1,6 +1,5 @@
 import React from "react";
-import PropTypes from "prop-types";
-import { connect } from "react-redux";
+import { useDispatch } from "react-redux";
 import styled from "styled-components";
 
 import { toggleMessageDetail } from "../../../actions/ui";
@@ -26,39 +25,30 @@ const ChatMenu = styled.div`
   }
 `;
 
-const ChatBar = ({ username, messages, dispatch }) => (
-  <ChatBarWrapper>
-    <h2>
-      {username}
-      {messages.length ? (
-        <em>
-          &nbsp; (<strong>{messages.length}</strong> messages)
-        </em>
-      ) : (
-        ""
-      )}
-    </h2>
-    <ChatMenu>
-      <Icon name="phone" style={{ margin: "0 0.5em" }} />
-      <Icon name="video" style={{ margin: "0 0.5em" }} />
-      <a onClick={() => dispatch(toggleMessageDetail())}>
-        <Icon name="info-circle" active style={{ margin: "0 0.5em" }} />
-      </a>
-    </ChatMenu>
-  </ChatBarWrapper>
-);
+const ChatBar = ({ username, messages }) => {
+  const dispatch = useDispatch();
 
-ChatBar.propTypes = {
-  messages: PropTypes.array.isRequired,
-  username: PropTypes.string.isRequired,
-  dispatch: PropTypes.func.isRequired
+  return (
+    <ChatBarWrapper>
+      <h2>
+        {username}
+        {messages.length ? (
+          <em>
+            &nbsp; (<strong>{messages.length}</strong> messages)
+          </em>
+        ) : (
+          ""
+        )}
+      </h2>
+      <ChatMenu>
+        <Icon name="phone" style={{ margin: "0 0.5em" }} />
+        <Icon name="video" style={{ margin: "0 0.5em" }} />
+        <a onClick={() => dispatch(toggleMessageDetail())}>
+          <Icon name="info-circle" active style={{ margin: "0 0.5em" }} />
+        </a>
+      </ChatMenu>
+    </ChatBarWrapper>
+  );
 };
 
-const mapDispatchToProps = dispatch => ({
-  dispatch
-});
-
-export default connect(
-  null,
-  mapDispatchToProps
-)(ChatBar);
+export default ChatBar;
